requirements.txt
Django>=5.1.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.5
drf-yasg>=1.21.5
reportlab>=4.0.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
python-dotenv==1.0.1
supabase-py==0.0.2 